requirements.txt
streamlit
fastapi
requests
uvicorn
matplotlib
langchain_community
langchain_openai
faiss-cpu
openai
grok
seaborn
plotly
pandas
torch
google-cloud-aiplatform
google-auth
vertexai